Moving twitter to dal, service will handle the data
